requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-socks==0.8.4
aiosignal==1.3.2
atomicwrites==1.4.1
attrs==25.1.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
docopt==0.6.2
frozenlist==1.5.0
h11==0.14.0
h2==4.1.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
matrix-nio==0.24.0
multidict==6.1.0
peewee==3.17.8
pillow==10.4.0
pipreqs==0.4.13
propcache==0.2.1
pycparser==2.22
pycryptodome==3.21.0
python-cryptography-fernet-wrapper==1.0.4
python-olm==3.2.16
python-socks==2.7.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
simplematrixbotlib==2.12.2
toml==0.10.2
unpaddedbase64==2.1.0
urllib3==2.3.0
yarg==0.1.10
yarl==1.18.3
